Move runtime_transcoder. Fix a bug in the UTF-16 decoder.
