+++ CMakeLists.txt
@@ -57,7 +57,6 @@ endif (ICUBABY_FUZZTEST)
 
 add_library (icubaby INTERFACE
   "${icubaby_project_root}/include/icubaby/icubaby.hpp"
-  "${icubaby_project_root}/include/icubaby/dy.hpp"
 )
 target_include_directories (icubaby INTERFACE "${icubaby_project_root}/include")
