+++ unittests/CMakeLists.txt
@@ -22,10 +22,10 @@
 
 add_executable (icubaby-unittests
   encoded_char.hpp
-  test_dy.cpp
-  test_u8.cpp
+  test_rtrans.cpp
   test_u16.cpp
   test_u32.cpp
+  test_u8.cpp
   test_utility.cpp
   typed_test.hpp
 )
